Unlocking classes
Modified the code for the main Player to simply "unlock" a class every time "A" is hit on the keyboard; hopefully setting the groundwork to making it so you unlock classes throughout the story as opposed to having them from the start.
Also made it so that in the top left, it draws icon sprites of each class as long as it's unlocked so you can actually see what classes you've set to be unlocked.